+++ com.onesignal.unity.android/Runtime/OneSignalAndroid.cs
@@ -107,6 +107,9 @@ namespace OneSignalSDK {
             set => _sdkClass.CallStatic("setRequiresUserPrivacyConsent", value);
         }
 
+        public override void SetLaunchURLsInApp(bool launchInApp)
+            => SDKDebug.Warn("This feature is only available for iOS.");
+
         public override void Initialize(string appId) {
             var unityPlayer = new AndroidJavaClass("com.unity3d.player.UnityPlayer");
             var activity    = unityPlayer.GetStatic<AndroidJavaObject>("currentActivity");